Combine NSIS plugins into a single mullvad_nsis.dll
Cheaper and simpler because we do not have to duplicate code across all
DLLs.

The namespace in installer.nsh is now `mullvad_nsis`.

This commit also includes a lot of refactoring work.
