+++ .github/workflows/mullvad-nsis.yml
@@ -25,13 +25,7 @@ jobs:
       - name: Build NSIS plugins
         shell: bash
         run: |
-          ci/cargo-ci.sh build --target i686-pc-windows-msvc \
-            -p mullvad-nsis \
-            -p nsis-registry \
-            -p nsis-pathedit \
-            -p nsis-cleanup \
-            -p nsis-log \
-            -p nsis-tray
+          ci/cargo-ci.sh build --target i686-pc-windows-msvc -p mullvad-nsis
 
       - name: Check working directory is clean
         shell: bash
