+++ Cargo.toml
@@ -18,11 +18,6 @@ members = [
   "mullvad-management-interface",
   "mullvad-masque-proxy",
   "mullvad-nsis",
-  "mullvad-nsis/plugins/cleanup",
-  "mullvad-nsis/plugins/log",
-  "mullvad-nsis/plugins/pathedit",
-  "mullvad-nsis/plugins/registry",
-  "mullvad-nsis/plugins/tray",
   "mullvad-paths",
   "mullvad-problem-report",
   "mullvad-relay-selector",
