--- mullvad-nsis/plugins/log/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "nsis-log"
-edition.workspace = true
-rust-version.workspace = true
-description = "NSIS log plugin - installer logging for the Mullvad VPN installer"
-repository.workspace = true
-license.workspace = true
-
-[lib]
-name = "log"
-crate-type = ["cdylib"]
-
-[target.i686-pc-windows-msvc.dependencies]
-nsis-plugin-api.workspace = true
-mullvad-nsis = { path = "../.." }
-mullvad-paths = { path = "../../../mullvad-paths" }
-talpid-platform-metadata = { path = "../../../talpid-platform-metadata" }
-windows-sys = { workspace = true, features = [
-  "Win32_Foundation",
-  "Win32_System_LibraryLoader",
-  "Win32_System_SystemInformation",
-  "Win32_UI_Input_KeyboardAndMouse",
-  "Win32_UI_WindowsAndMessaging",
-] }
-
-[lints]
-workspace = true
