--- mullvad-nsis/plugins/registry/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "nsis-registry"
-edition.workspace = true
-rust-version.workspace = true
-description = "NSIS registry plugin - registry operations for the Mullvad VPN installer"
-repository.workspace = true
-license.workspace = true
-
-[lib]
-name = "registry"
-crate-type = ["cdylib"]
-
-[target.i686-pc-windows-msvc.dependencies]
-nsis-plugin-api.workspace = true
-windows-sys = { workspace = true, features = [
-  "Win32_Foundation",
-  "Win32_System_Registry",
-] }
-
-[lints]
-workspace = true
