--- mullvad-nsis/plugins/tray/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "nsis-tray"
-edition.workspace = true
-rust-version.workspace = true
-description = "NSIS tray plugin - tray icon management for the Mullvad VPN installer"
-repository.workspace = true
-license.workspace = true
-
-[lib]
-name = "tray"
-crate-type = ["cdylib"]
-
-[target.i686-pc-windows-msvc.dependencies]
-nsis-plugin-api.workspace = true
-windows-sys = { workspace = true, features = [
-  "Win32_Foundation",
-  "Win32_Security",
-  "Win32_System_Com",
-  "Win32_System_ProcessStatus",
-  "Win32_System_Registry",
-  "Win32_System_SystemInformation",
-  "Win32_System_Threading",
-  "Win32_System_Time",
-  "Win32_UI_Shell",
-] }
-
-[lints]
-workspace = true
