+++ mullvad-nsis/src/tray/mullvad_tray_record.bin
